+++ src/main/java/com/myboard/firebase/fcm/FCMService.java
@@ -4,18 +4,10 @@ import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.Message;
 import com.google.firebase.messaging.WebpushConfig;
 import com.google.firebase.messaging.WebpushNotification;
-import com.myboard.entity.Article;
-import com.myboard.entity.User;
-import com.myboard.repository.article.ArticleRepository;
-import com.myboard.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
-import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 
 @Slf4j
@@ -23,20 +15,7 @@ import java.util.concurrent.ExecutionException;
 @RequiredArgsConstructor
 public class FCMService {
 
-    private final FCMTokenManager fcmTokenManager;
-    private final ArticleRepository articleRepository;
-    private final UserRepository userRepository;
-
-    @Value("${firebase.notifications.defaults.title}")
-    private String defaultTitle;
-
-    @Value("${firebase.notifications.defaults.message}")
-    private String defaultMessage;
-
-    @Value("${firebase.notifications.defaults.token}")
-    private String defaultToken;
-
-    private void sendMessage(String token, String title, String contents) {
+    public void sendMessage(String token, String title, String contents) {
         // setToken 혹은 setTopic을 이용해 메세지의 타겟을 결정
         Message message = Message.builder()
                 .setToken(token)
@@ -53,31 +32,5 @@ public class FCMService {
         } catch (ExecutionException | InterruptedException e) {
             throw new IllegalStateException("알림 전송에 실패하였습니다.");
         }
-    }
-
-    // 게시글에 댓글이 달리면 알림 전송
-    @Async
-    public void commentNotification(Long articleId, Long commentAuthorId) {
-        Optional<Article> targetArticle = articleRepository.findArticleByIdFetchJoin(articleId);
-        Optional<User> commentAuthor = userRepository.findById(commentAuthorId);
-
-        if (targetArticle.isPresent() && commentAuthor.isPresent()) {
-            Long targetUserId = targetArticle.get().getUser().getId();
-            String targetArticleTitle = targetArticle.get().getTitle();
-            String targetUsername = targetArticle.get().getUser().getUsername();
-            String commentAuthorName = commentAuthor.get().getUsername();
-            String targetUserToken = fcmTokenManager.getToken(String.valueOf(targetUserId));
-
-            Optional.ofNullable(targetUserToken).ifPresent(token -> sendMessage(
-                    token,
-                    "새로운 댓글이 작성되었습니다!",
-                    targetUsername + "님의" + targetArticleTitle + "게시글에" + commentAuthorName + "님이 댓글을 남겼습니다.")
-            );
-        }
-    }
-
-    @Async
-    public void testNotification() {
-        sendMessage(defaultToken ,defaultTitle, defaultMessage);
     }
 }
