+++ src/main/java/com/myboard/service/articleComment/ArticleCommentServiceImpl.java
@@ -6,7 +6,7 @@ import com.myboard.entity.Article;
 import com.myboard.entity.ArticleComment;
 import com.myboard.entity.User;
 import com.myboard.exception.user.NotAuthorException;
-import com.myboard.firebase.fcm.FCMService;
+import com.myboard.firebase.fcm.PushNotificationService;
 import com.myboard.repository.article.ArticleRepository;
 import com.myboard.repository.articleComment.ArticleCommentRepository;
 import com.myboard.repository.user.UserRepository;
@@ -23,7 +23,7 @@ public class ArticleCommentServiceImpl implements ArticleCommentService{
     private final UserRepository userRepository;
     private final ArticleRepository articleRepository;
     private final ArticleCommentRepository articleCommentRepository;
-    private final FCMService fcmService;
+    private final PushNotificationService pushNotificationService;
 
     @Override
     @Transactional
@@ -40,7 +40,7 @@ public class ArticleCommentServiceImpl implements ArticleCommentService{
         ArticleComment savedArticleComment = articleCommentRepository.save(articleComment);
 
         log.info("Fcm start");
-        fcmService.commentNotification(articleId, userId);
+        pushNotificationService.commentNotification(articleId, userId);
         log.info("Fcm finish");
 
         log.info("return ArticleCommentId");
